paypal

+++ order-service/src/main/java/com/keiko/orderservice/entity/resources/Price.java
@@ -4,10 +4,12 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 public class Price {
-    private Double value;
+    private BigDecimal value;
     private boolean isPromotional;
 }
